Move blocks to separate module
