+++ src/block.rs
@@ -1,0 +1,11 @@
+pub mod ffm;
+pub mod iterators;
+pub mod loss_functions;
+pub mod lr;
+pub mod misc;
+pub mod neural;
+pub mod normalize;
+pub mod relu;
+#[cfg(test)]
+pub mod test;
+pub mod file;
