--- src/version.rs
@@ -1,1 +1,0 @@
-pub static LATEST: &str = "0.2";
